requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
fastmcp
pydantic>=2.6.0